Correções path
Refactor dos paths após a estruturação das pastas
