+++ src/modules/Users/FindAllUsers/FindAllUsersRepository.ts
@@ -1,5 +1,5 @@
 import User from "@models/User";
-import portal from "src/database/prisma/portalClient";
+import portal from "@database/prisma/portalClient";
 
 export interface IFindAllUsersRepository {
   findAll(): Promise<User[]>;
