+++ tsconfig.json
@@ -26,14 +26,23 @@
             "@configs/*": [
                 "./src/configs/*"
             ],
-            "@controllers/*": [
-                "./src/controllers/*"
+            "@database/*": [
+                "./src/database/*"
             ],
             "@errors/*": [
                 "./src/errors/*"
             ],
             "@models/*": [
                 "./src/models/*"
+            ],
+            "@modules/*": [
+                "./src/modules/*"
+            ],
+            "@routes/*": [
+                "./src/routes/*"
+            ],
+            "@static/*": [
+                "./src/static/*"
             ]
         }
     },
